extend shape calculation for preprocessings
