+++ tools/accuracy_checker/accuracy_checker/preprocessor/__init__.py
@@ -28,16 +28,12 @@ from .color_space_conversion import (
     BGRToLAB, RGBToLAB
 )
 from .audio_preprocessing import (
-    ResampleAudio,
-    ClipAudio,
-    NormalizeAudio,
     HanningWindow,
     AudioSpectrogram,
     TriangleFiltering,
     DCT,
     ClipCepstrum,
     TrimmingAudio,
-    SamplesToFloat32,
     AudioToMelSpectrogram
 )
 from .audio_preprocessing_ext import (
@@ -46,7 +42,11 @@ from .audio_preprocessing_ext import (
     DitherSpectrum,
     PreemphFrame,
     SignalPatching,
-    ContextWindow
+    ContextWindow,
+    ResampleAudio,
+    ClipAudio,
+    SamplesToFloat32,
+    NormalizeAudio
 )
 
 from .normalization import Normalize, Normalize3d
